# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  server.apiResponse:
    properties:
      data: {}
      meta: {}
    type: object
...
info:
  contact: {}
  description: 'Backend service for the NEXQIA Healthcare Operating Platform. All
    routes are under /api/v1. Response envelope: {"data": ..., "meta": {...}}.'
  title: NEXQIA API
  version: "1.0"
paths:
...
  /display/queue:
    get:
      description: Read-only, no real-time push in v1.
      parameters:
      - description: Active merchant UUID
        in: header
        name: X-Merchant-ID
        required: true
        type: string
      - description: Filter by department UUID
        in: query
        name: department_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.apiResponse'
      security:
      - BearerAuth: []
      summary: Poll-based queue board for a merchant
      tags:
      - display
  /display/schedule:
    get:
      parameters:
      - description: Active merchant UUID
        in: header
        name: X-Merchant-ID
        required: true
        type: string
      - description: Department UUID
        in: query
        name: department_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.apiResponse'
      security:
      - BearerAuth: []
      summary: Today's active physician schedules for a department
      tags:
      - display
...
swagger: "2.0"